embed plan generation steps
