+++ worker.js
@@ -34,7 +34,64 @@ async function sendEmailUniversal(to, subject, body, env = {}) {
   await sendEmail(to, subject, body, phpEnv);
 }
 import { parseJsonSafe } from './utils/parseJsonSafe.js';
-import { generateProfile, generateMenu, generatePrinciples, generateGuidance } from "./worker_modules/planGenerationSteps.js";
+
+// ----- START GENERATED PLAN STEP FUNCTIONS -----
+async function generateProfile(replacements, env, modelName, userId) {
+  const template = await env.RESOURCES_KV.get('prompt_generate_profile');
+  if (!template) throw new Error('Missing prompt_generate_profile');
+  const populated = populatePrompt(template, replacements);
+  const raw = await callModel(modelName, populated, env, { temperature: 0.1, maxTokens: 4000 });
+  const cleaned = cleanGeminiJson(raw);
+  const parsed = safeParseJson(cleaned, {});
+  await env.USER_METADATA_KV.put(
+    `plan_section_profile_${userId}`,
+    JSON.stringify({ ts: Date.now(), data: parsed })
+  );
+  return parsed;
+}
+
+async function generateMenu(replacements, env, modelName, userId) {
+  const template = await env.RESOURCES_KV.get('prompt_generate_menu');
+  if (!template) throw new Error('Missing prompt_generate_menu');
+  const populated = populatePrompt(template, replacements);
+  const raw = await callModel(modelName, populated, env, { temperature: 0.1, maxTokens: 4000 });
+  const cleaned = cleanGeminiJson(raw);
+  const parsed = safeParseJson(cleaned, {});
+  await env.USER_METADATA_KV.put(
+    `plan_section_menu_${userId}`,
+    JSON.stringify({ ts: Date.now(), data: parsed })
+  );
+  return parsed;
+}
+
+async function generatePrinciples(replacements, env, modelName, userId) {
+  const template = await env.RESOURCES_KV.get('prompt_generate_principles');
+  if (!template) throw new Error('Missing prompt_generate_principles');
+  const populated = populatePrompt(template, replacements);
+  const raw = await callModel(modelName, populated, env, { temperature: 0.1, maxTokens: 4000 });
+  const cleaned = cleanGeminiJson(raw);
+  const parsed = safeParseJson(cleaned, {});
+  await env.USER_METADATA_KV.put(
+    `plan_section_principles_${userId}`,
+    JSON.stringify({ ts: Date.now(), data: parsed })
+  );
+  return parsed;
+}
+
+async function generateGuidance(replacements, env, modelName, userId) {
+  const template = await env.RESOURCES_KV.get('prompt_generate_guidance');
+  if (!template) throw new Error('Missing prompt_generate_guidance');
+  const populated = populatePrompt(template, replacements);
+  const raw = await callModel(modelName, populated, env, { temperature: 0.1, maxTokens: 4000 });
+  const cleaned = cleanGeminiJson(raw);
+  const parsed = safeParseJson(cleaned, {});
+  await env.USER_METADATA_KV.put(
+    `plan_section_guidance_${userId}`,
+    JSON.stringify({ ts: Date.now(), data: parsed })
+  );
+  return parsed;
+}
+// ----- END GENERATED PLAN STEP FUNCTIONS -----
 
 const WELCOME_SUBJECT = 'Добре дошъл в MyBody!';
 const WELCOME_BODY_TEMPLATE = `<!DOCTYPE html>
